+++ kong/globalpatches.lua
