+++ kong/plugins/response-ratelimiting/policies/init.lua
@@ -2,7 +2,7 @@ local singletons = require "kong.singletons"
 local timestamp = require "kong.tools.timestamp"
 local redis = require "resty.redis"
 local policy_cluster = require "kong.plugins.response-ratelimiting.policies.cluster"
-local reports = require "kong.core.reports"
+local reports = require "kong.reports"
 
 
 local ngx_log = ngx.log
